src/function_orchestrate_ingest/requirements.txt
anyio==4.6.2.post1
asyncpg==0.30.0
azure-core==1.32.0
azure-eventhub==5.13.0
azure-identity==1.19.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-storage-blob==12.24.0
certifi==2024.8.30 
cffi==1.17.1 
charset-normalizer==3.4.0
cryptography==43.0.3 
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6 
httpx==0.27.2
idna==3.10
isodate==0.7.2 
numpy==2.1.3
pandas==2.2.3
pycparser==2.22 
python-dateutil==2.9.0.post0 
psycopg2-binary==2.9.10
pytz==2024.2 
requests==2.32.3 
six==1.16.0 
sniffio==1.3.1 
sqlalchemy==2.0.36
typing-extensions==4.12.2 
tzdata==2024.2 
urllib3==2.2.3 
